L2: converted global indices to long

+++ src/L2-foundation/BoSSS.Foundation.XDG/MultiphaseAgglomerator.cs
@@ -275,7 +275,7 @@ namespace BoSSS.Foundation.XDG {
                 }
             }
 
-            public int i0Func(int jCell, int iVar) {
+            public long i0Func(int jCell, int iVar) {
                 if (VarIsXdg[iVar]) {
                     int iSpc = m_LsRegion.GetSpeciesIndex(this.m_spId, jCell);
                     return m_Map.GlobalUniqueCoordinateIndex(iVar, jCell, iSpc * NS[iVar]);
@@ -780,7 +780,7 @@ namespace BoSSS.Foundation.XDG {
                 int Jup = grdDat.Cells.NoOfLocalUpdatedCells;
                 int Jtot = grdDat.Cells.Count;
                 Partitioning CellPart = Tracker.GridDat.CellPartitioning;
-                int i0 = CellPart.i0;
+                long i0 = CellPart.i0;
                 var GidxExt = Tracker.GridDat.Parallel.GlobalIndicesExternalCells;
                 var GidxExt2Lidx = Tracker.GridDat.Parallel.Global2LocalIdx;
                 //double[] RefVolumes = grdDat.Grid.RefElements.Select(Kref => Kref.Volume).ToArray();
